Beatitude allows the user to play bass live over the drum sequences, compiles with latest version (V1.8.5) of Arduino IDE
1) The new version of Beatitude allows the user to play bass live over
the drum sequences -- i.e. when a drum sequence is being played back by
the ArduTouch (the red LED is flashing) then the keyboard becomes a bass
sound that can be played spontaneously over the drum track.

2) The library now provides "lofi" versions of all the drum samples. The
example sketch _30_DrumPad.ino contains a discussion of these.

3) Sawtooth and square wave oscillators have been added  in a new
library module WaveGen. The waveforms they generate are computed on the
fly, they are not wave-table driven. You've met the sawtooth before  in
Thick. The square wave oscillator is new. It features a pulse width
parameter, an internal LFO that can modulate the pulse width (around the
aforementioned parameter as a center pt), and an internal low pass
filter. The pw LFO really thickens up the oscillator -- the bass sound
used by Beatitude is generated by only 2 square wave oscillators
(whereas Thick uses 4 sawtooth oscillators.)

4) There is a BUILD_166 define in Model.h (which is commented out by
default). It is important to uncomment this line if you are using 1.6.6.

5) The identifier "synth" is now a reserved symbol of the ArduTouch
library. It provides a backptr to the synth object which is instantiated
in the call to ardutouch_setup(), and is declared extern so that all
objects in the system  can directly call the main synth object (whatever
it may be). Most of the sketches that had been written up to now had
used "synth" casually here and there. All of these instances had to be
changed ( ...arghh).

6)  Beatitude can now be successfully compiled under the latest Arduino
builds. The cost is a lofi version of the snare sound, which is somewhat
noticeable.  But both the 1.6.6 and 1.8.6 versions of Beatitude use lofi
versions of the bass and tom sounds and these are almost
indistinguishable from their normal counterparts :)