New Mantra synth; Library has several new base classes
New Mantra synthesizer, which plays a few layers of Indian-like drone
music  (choice of two drones), with an exotic scale you can play on top
of it, with an inability to make any music that doesn't sound good.  :)

The library  now includes several new base classes, among which are:

FiltEnv              - a dynamic (i.e. enveloped) low-pass filter
DelayVibrato   -  adds a delay time before triggering stock vibrato
DualOsc           -  allows for the transparent pairing of any two
oscillators.

Additionally, the old Sequencer class has been re-written, and is now
abstract in that it
doesn't know where the compiled sequence is even stored. Then the
following 2
intermediate sequencer classes have been added:

SequencerRAM - a Sequencer whose sequence resides in RAM
SequencerROM - a Sequencer whose sequence resides in ROM

and the following high level class has been added:

RealTimeSequencer  - imported from  Beatitude synth, then changed so
that
it derives from the new SequencerRAM class.

The abstraction layer introduced in the Sequencer class is important
because
it allows one, for example, to derive from Sequencer a new class in
which the sequence does not reside in RAM or ROM or anywhere, but is
produced algorithmically on the fly.