Updated Beatitude, bug fixes in RealTimeSequencer in library, new demo video
=== RELEASE 1.14 ===

This release is all about Beatitude and the RealTimeSequencer class  in the library.

The new version of Beatitude has these added features:
   1. The drum kit is now panned in stereo.
   2. The top pot controls the balance between the drum kit and the bass.
   3. When recording a sequence with prior versions of Beatitude the 1st
      drum hit was often missed if it happened "right before" the 1st beat
      of the 1st measure. This has been fixed.
   4. Fixed problems with sequence playback via changes to the RealTimeSequencer
         class in the library (see below).

RealTimeSequencer::jiffsToTrans is relaxed to the duration of a sixteenth note (was 1 juff)
         the following fixes to the RealTimeSequencer class ...
         bugfix: super was type-defed as Sequencer instead of SequencerRAM!
         bugfix: multiple key events during transition no longer cause haywire record compilation
         bugfix: :evHandler() no longer compiles bogus rest records
         bugfix: when dynamics() concatenated a rest it was not checking for an empty buffer

The last release (1.13) fixed the long-running problem in both Beatitude and Mantra
which caused both to behave bizarrely at times depending on what prior sketch had
 been loaded in the ROM. In the case of Beatitude it turns out that this problem (which
 was caused by a routine being called from the constructor of a static instantiation of the Beatitude synth before the library had been initialized -- and on which it depended) was
masking other serious issues: this new version of the library fixes not 1, not 2, not 3,
but 4 different bugs involving  corner cases in the RealTimeSequencer class.

The only items that are updated in this relase are: the Beatitude synth and the
library -- everything else remains the same.

I also added a video demo of Mitch playing 7 ArduTouch synths at once.